+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased]
+
+### Added
+* Added background blur and background replacement features. More details [here](https://github.com/aws/amazon-chime-sdk-android/blob/master/guides/background_video_filters.md).
+
 ## [0.20.0] - 2022-05-11
 
 ### Added